Break circular chain
