+++ packages/kbn-reporting/common/index.ts
@@ -10,7 +10,6 @@ export * from './constants';
 export * from './errors';
 export * from './schema_utils';
 
-export { jobTypes } from './job_types';
 export { buildKibanaPath } from './build_kibana_path';
 export { CancellationToken } from './cancellation_token';
 export { getFullRedirectAppUrl } from './get_full_redirect_app_url';
