+++ x-pack/plugins/reporting/public/share_context_menu/register_csv_reporting.tsx
@@ -7,11 +7,13 @@
 
 import { i18n } from '@kbn/i18n';
 import React from 'react';
+
+import { CSV_JOB_TYPE } from '@kbn/reporting-export-types-csv-common';
+
 import type { SearchSourceFields } from '@kbn/data-plugin/common';
 import { ShareContext, ShareMenuProvider } from '@kbn/share-plugin/public';
-import { CSV_JOB_TYPE } from '@kbn/reporting-common';
 import { checkLicense } from '../lib/license_check';
-import { ExportPanelShareOpts } from '.';
+import type { ExportPanelShareOpts } from '.';
 import { ReportingPanelContent } from './reporting_panel_content_lazy';
 
 export const reportingCsvShareProvider = ({
