+++ x-pack/plugins/reporting/public/shared/get_shared_components.tsx
@@ -6,8 +6,11 @@
  */
 
 import { CoreSetup } from '@kbn/core/public';
-import { PDF_REPORT_TYPE_V2, PDF_REPORT_TYPE, PNG_REPORT_TYPE_V2 } from '@kbn/reporting-common';
 import React from 'react';
+
+import { PDF_REPORT_TYPE, PDF_REPORT_TYPE_V2 } from '@kbn/reporting-export-types-pdf-common';
+import { PNG_REPORT_TYPE_V2 } from '@kbn/reporting-export-types-png-common';
+
 import { ReportingAPIClient } from '../lib/reporting_api_client';
 import { ReportingPanelProps } from '../share_context_menu/reporting_panel_content';
 import { ScreenCapturePanelContent } from '../share_context_menu/screen_capture_panel_content_lazy';
