+++ x-pack/plugins/reporting/server/lib/check_params_version.ts
@@ -6,8 +6,8 @@
  */
 
 import type { Logger } from '@kbn/core/server';
+import { BaseParams } from '@kbn/reporting-common/types';
 import { UNVERSIONED_VERSION } from '../../common/constants';
-import type { BaseParams } from '../types';
 
 export function checkParamsVersion(jobParams: BaseParams, logger: Logger) {
   if (jobParams.version) {
