+++ x-pack/plugins/reporting/server/routes/common/jobs/get_document_payload.test.ts
@@ -5,8 +5,11 @@
  * 2.0.
  */
 
-import { PDF_JOB_TYPE, CSV_JOB_TYPE, PDF_JOB_TYPE_V2 } from '@kbn/reporting-common';
 import { Readable } from 'stream';
+
+import { CSV_JOB_TYPE } from '@kbn/reporting-export-types-csv-common';
+import { PDF_JOB_TYPE, PDF_JOB_TYPE_V2 } from '@kbn/reporting-export-types-pdf-common';
+
 import { ReportApiJSON } from '../../../../common/types';
 import { ContentStream, getContentStream, statuses } from '../../../lib';
 import { createMockConfigSchema, createMockReportingCore } from '../../../test_helpers';
